requirements.txt
aiogram
google-generativeai
apscheduler
aiosqlite
python-dotenv
